Handle KeyStoreException in Android 8.0
This exception is thrown in "verify" on Android 8.0 if new fingerprints were enrolled after "save" has been called. On Android 8.1 KeyPermanentlyInvalidatedException is thrown in the same scenario (but already in "has"), therefore I'm handling it the same - i.e. deleting the key so that it will be recreated on next call.

When the application receives this error (KeyPermanentlyInvalidatedException), it should reinitialize the fingerprint usage - the user will have to reenter the password and agree to store it using the fingerprint.